Restructure files and directories
Restructure the project with hexagonal architectural approach

Its main motivation is to separate our application into different layers
or regions with its own responsibility. In this way it manages to decouple
layers of our application allowing them to evolve in isolation. In addition,
having the system separated by responsibilities will facilitate reuse.
