+++ src/application/services/CarInsurance.js
@@ -1,36 +1,27 @@
-class Product {
-  constructor(name, sellIn, price) {
-    this.name = name;
-    this.sellIn = sellIn;
-    this.price = price;
-  }
-}
-
 class CarInsurance {
   constructor(products = []) {
     this.products = products;
   }
+
   updatePrice() {
-    for (var i = 0; i < this.products.length; i++) {
+    for (let i = 0; i < this.products.length; i++) {
       if (this.products[i].name != 'Full Coverage' && this.products[i].name != 'Special Full Coverage') {
         if (this.products[i].price > 0) {
           if (this.products[i].name != 'Mega Coverage') {
             this.products[i].price = this.products[i].price - 1;
           }
         }
-      } else {
-        if (this.products[i].price < 50) {
-          this.products[i].price = this.products[i].price + 1;
-          if (this.products[i].name == 'Special Full Coverage') {
-            if (this.products[i].sellIn < 11) {
-              if (this.products[i].price < 50) {
-                this.products[i].price = this.products[i].price + 1;
-              }
+      } else if (this.products[i].price < 50) {
+        this.products[i].price = this.products[i].price + 1;
+        if (this.products[i].name == 'Special Full Coverage') {
+          if (this.products[i].sellIn < 11) {
+            if (this.products[i].price < 50) {
+              this.products[i].price = this.products[i].price + 1;
             }
-            if (this.products[i].sellIn < 6) {
-              if (this.products[i].price < 50) {
-                this.products[i].price = this.products[i].price + 1;
-              }
+          }
+          if (this.products[i].sellIn < 6) {
+            if (this.products[i].price < 50) {
+              this.products[i].price = this.products[i].price + 1;
             }
           }
         }
@@ -49,10 +40,8 @@ class CarInsurance {
           } else {
             this.products[i].price = this.products[i].price - this.products[i].price;
           }
-        } else {
-          if (this.products[i].price < 50) {
-            this.products[i].price = this.products[i].price + 1;
-          }
+        } else if (this.products[i].price < 50) {
+          this.products[i].price = this.products[i].price + 1;
         }
       }
     }
@@ -61,7 +50,4 @@ class CarInsurance {
   }
 }
 
-module.exports = {
-  Product,
-  CarInsurance
-}
+module.exports = CarInsurance;
