+++ src/domain/model/Product.js
@@ -1,0 +1,9 @@
+class Product {
+  constructor(name, sellIn, price) {
+    this.name = name;
+    this.sellIn = sellIn;
+    this.price = price;
+  }
+}
+
+module.exports = Product;
