+++ test/coTestSpec.js
@@ -1,14 +1,12 @@
 /* eslint-disable prefer-arrow-callback */
 const { expect } = require('chai');
 
-const coTest = require('../src/coTest');
-
-const { CarInsurance } = coTest;
-const { Product } = coTest;
+const CarInsurance = require('../src/application/services/CarInsurance');
+const Product = require('../src/domain/model/Product');
 
 describe('Co Test', function () {
   context('when name is Medium Coverage', function () {
-    const data = require('./medium-coverage-data.json');
+    const data = require('./assets/medium-coverage-data.json');
 
     const { name } = data;
     const initialSellIn = data.initialValues.sellIn;
@@ -28,7 +26,7 @@ describe('Co Test', function () {
   });
 
   context('when name is Full Coverage', function () {
-    const data = require('./full-coverage-data.json');
+    const data = require('./assets/full-coverage-data.json');
 
     const { name } = data;
     const initialSellIn = data.initialValues.sellIn;
@@ -48,7 +46,7 @@ describe('Co Test', function () {
   });
 
   context('when name is Low Coverage', function () {
-    const data = require('./low-coverage-data.json');
+    const data = require('./assets/low-coverage-data.json');
 
     const { name } = data;
     const initialSellIn = data.initialValues.sellIn;
